+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.12.2
+* Add lex2 example - add kdb-charts on the fly.
+* fix innerHTML issue when adding new data.
+* fix errors with updates via kd-query dispatch.
+
 ## 0.12.1
 * Add .z.wc to srv.q